refactor: Clean up
